requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
sqlalchemy==2.0.30
langchain==0.2.1
langchain-groq==0.1.4
langchain-google-genai==1.0.5
tavily-python==0.3.3
python-dotenv==1.0.1
